chore(team): split run creation manager tests
